Don't limit the context to a certain output medium
This kind of information should not have any html and also should not be echoed in this layer of the application. That way you couple the use of this class to an environment in which you're able to echo some information and in which html is processed, so the proper use of the exception case is very limited.